requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
aiogram==3.13.1
python-dotenv==1.0.1
python-slugify==8.0.4
httpx==0.27.2
openai==1.44.0
python-multipart==0.0.9
pytest==8.3.2
